+++ src/usr.sbin/ppp/filter.c
@@ -47,7 +47,7 @@ struct filterent afilters[MAXFILTERS];	/* keep-alive packet filter */
 
 static struct filterent filterdata;
 
-static u_long netmasks[33] = {
+static u_int32_t netmasks[33] = {
   0x00000000,
   0x80000000, 0xC0000000, 0xE0000000, 0xF0000000,
   0xF8000000, 0xFC000000, 0xFE000000, 0xFF000000,
